KB-10909 | BE | DEV | Enhancement in the create post API to check the profanity.
1. Refractored the code and moved as a seperate module.
