+++ src/test/java/com/igot/cb/profanity/consumer/ProfanityConsumerTest.java
@@ -1,4 +1,4 @@
-package com.igot.cb.consumer;
+package com.igot.cb.profanity.consumer;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -6,7 +6,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.igot.cb.discussion.repository.DiscussionRepository;
 import com.igot.cb.pores.util.Constants;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.*;
