+++ src/test/java/com/igot/cb/service/impl/DiscussionServiceImplTest.java
@@ -3364,74 +3364,4 @@ class DiscussionServiceImplTest {
         assertEquals("Index: 0, Size: 0",response.getParams().getErrMsg());
     }
 
-
-    @Test
-    void testProcessProfanityCheck_HappyPath() throws Exception {
-        UUID id = UUID.randomUUID();
-        ObjectNode node = JsonNodeFactory.instance.objectNode();
-        node.put(Constants.DESCRIPTION, "desc");
-        node.put(Constants.LANGUAGE_CODE, "en");
-        when(cbServerProperties.getCbDiscussionApiKey()).thenReturn("api-key");
-        when(cbServerProperties.getCbServiceRegistryBaseUrl()).thenReturn("http://base-url");
-        when(cbServerProperties.getCbRegistryTextModerationApiPath()).thenReturn("moderation");
-        Method m = DiscussionServiceImpl.class.getDeclaredMethod("processProfanityCheck", UUID.class, ObjectNode.class);
-        m.setAccessible(true);
-        m.invoke(discussionService, id, node);
-        verify(requestHandlerService).fetchResultUsingPost(anyString(), anyMap(), anyMap());
-    }
-
-    @Test
-    void testProcessProfanityCheck_MissingDescription() throws Exception {
-        UUID id = UUID.randomUUID();
-        ObjectNode node = JsonNodeFactory.instance.objectNode();
-        node.put(Constants.LANGUAGE_CODE, "en");
-        when(cbServerProperties.getCbDiscussionApiKey()).thenReturn("api-key");
-        when(cbServerProperties.getCbServiceRegistryBaseUrl()).thenReturn("http://base-url");
-        when(cbServerProperties.getCbRegistryTextModerationApiPath()).thenReturn("moderation");
-        Method m = DiscussionServiceImpl.class.getDeclaredMethod("processProfanityCheck", UUID.class, ObjectNode.class);
-        m.setAccessible(true);
-        InvocationTargetException ex = assertThrows(InvocationTargetException.class, () -> m.invoke(discussionService, id, node));
-        assertTrue(ex.getCause() instanceof NullPointerException);
-    }
-
-    @Test
-    void testProcessProfanityCheck_MissingLanguageCode() throws Exception {
-        UUID id = UUID.randomUUID();
-        ObjectNode node = JsonNodeFactory.instance.objectNode();
-        node.put(Constants.DESCRIPTION, "desc");
-        when(cbServerProperties.getCbDiscussionApiKey()).thenReturn("api-key");
-        when(cbServerProperties.getCbServiceRegistryBaseUrl()).thenReturn("http://base-url");
-        when(cbServerProperties.getCbRegistryTextModerationApiPath()).thenReturn("moderation");
-        Method m = DiscussionServiceImpl.class.getDeclaredMethod("processProfanityCheck", UUID.class, ObjectNode.class);
-        m.setAccessible(true);
-        InvocationTargetException ex = assertThrows(InvocationTargetException.class, () -> m.invoke(discussionService, id, node));
-        assertTrue(ex.getCause() instanceof NullPointerException);
-    }
-
-    @Test
-    void testProcessProfanityCheck_NullDiscussionDetailsNode() throws Exception {
-        UUID id = UUID.randomUUID();
-        when(cbServerProperties.getCbDiscussionApiKey()).thenReturn("api-key");
-        when(cbServerProperties.getCbServiceRegistryBaseUrl()).thenReturn("http://base-url");
-        when(cbServerProperties.getCbRegistryTextModerationApiPath()).thenReturn("moderation");
-        Method m = DiscussionServiceImpl.class.getDeclaredMethod("processProfanityCheck", UUID.class, ObjectNode.class);
-        m.setAccessible(true);
-        InvocationTargetException ex = assertThrows(InvocationTargetException.class, () -> m.invoke(discussionService, id, null));
-        assertTrue(ex.getCause() instanceof NullPointerException);
-    }
-
-    @Test
-    void testProcessProfanityCheck_NullId() throws Exception {
-        ObjectNode node = JsonNodeFactory.instance.objectNode();
-        node.put(Constants.DESCRIPTION, "desc");
-        node.put(Constants.LANGUAGE_CODE, "en");
-        when(cbServerProperties.getCbDiscussionApiKey()).thenReturn("api-key");
-        when(cbServerProperties.getCbServiceRegistryBaseUrl()).thenReturn("http://base-url");
-        when(cbServerProperties.getCbRegistryTextModerationApiPath()).thenReturn("moderation");
-        Method m = DiscussionServiceImpl.class.getDeclaredMethod("processProfanityCheck", UUID.class, ObjectNode.class);
-        m.setAccessible(true);
-        m.invoke(discussionService, null, node);
-        verify(requestHandlerService).fetchResultUsingPost(anyString(), anyMap(), anyMap());
-    }
-
 }
